avengers/requirement.txt
appdirs==1.4.3
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
cssselect==1.1.0
Django==2.2.5
django-ckeditor==5.9.0
django-js-asset==1.2.2
djangorestframework==3.10.2
fake-useragent==0.1.11
html2text==2019.8.11
idna==2.8
jieba==0.39
lxml==4.4.1
parse==1.12.1
pyee==6.0.0
pyppeteer==0.0.25
pyquery==1.4.0
pytz==2019.2
PyYAML==5.3
requests==2.22.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
tqdm==4.35.0
urllib3==1.25.3
w3lib==1.21.0
websockets==8.0.2
